+++ packages/appium/test/e2e/driver.e2e.spec.js
@@ -4,9 +4,9 @@ import {BaseDriver} from '@appium/base-driver';
 import {exec} from 'teen_process';
 import {fs, tempDir} from '@appium/support';
 import axios from 'axios';
-import {command} from 'webdriver';
 import B from 'bluebird';
 import _ from 'lodash';
+// eslint-disable-next-line import/named
 import {createSandbox} from 'sinon';
 import {remote as wdio} from 'webdriverio';
 import {runExtensionCommand} from '../../lib/cli/extension';
@@ -110,12 +110,14 @@ describe('FakeDriver via HTTP', function () {
    * @param {Partial<import('appium/types').ParsedArgs>} [args]
    */
   function withServer(args = {}) {
+    // eslint-disable-next-line mocha/no-sibling-hooks
     before(async function () {
       args = {...args, appiumHome, port, address: TEST_HOST};
       if (shouldStartServer) {
         server = await appiumServer(args);
       }
     });
+    // eslint-disable-next-line mocha/no-sibling-hooks
     after(async function () {
       if (server) {
         await server.close();
@@ -251,8 +253,6 @@ describe('FakeDriver via HTTP', function () {
 
       await B.delay(250);
       await driver.getPageSource().should.eventually.be.rejectedWith(/terminated/);
-
-      await driver.getSessions().should.eventually.be.empty;
     });
 
     it('should not allow umbrella commands to prevent newCommandTimeout on inner driver', async function () {
@@ -264,6 +264,10 @@ describe('FakeDriver via HTTP', function () {
       );
       let driver = await wdio({...wdOpts, capabilities: localCaps});
       should.exist(driver.sessionId);
+      driver.addCommand(
+        'getSessions',
+        async () => (await axios.get(`${testServerBaseUrl}/sessions`)).data.value
+      );
 
       // get the session list 6 times over 300ms. each request will be below the new command
       // timeout but since they are not received by the driver the session should still time out
@@ -506,51 +510,12 @@ describe('FakeDriver via HTTP', function () {
         await driver.deleteSession();
       }
     });
-
-    it.skip('should log a single deprecation warning if a deprecated method is used and not overridden by a newMethodMap', async function () {
-      let driver = await wdio({...wdOpts, capabilities: caps});
-      try {
-        driver.addCommand(
-          'deprecated',
-          command('POST', '/session/:sessionId/deprecated', {
-            command: 'deprecated',
-            description: 'Call a deprecated command',
-            parameters: [],
-            ref: '',
-          }),
-        );
-        driver.addCommand(
-          'doubleClick',
-          command('POST', '/session/:sessionId/doubleclick', {
-            command: 'doubleClick',
-            description: 'Global double click',
-            parameters: [],
-            ref: '',
-          }),
-        );
-        await driver
-          .executeScript('fake: getDeprecatedCommandsCalled', [])
-          .should.eventually.eql([]);
-        await driver.deprecated();
-        await driver.deprecated();
-        await driver.shake();
-
-        // this call should not trigger a deprecation even though deprecated by appium because it's
-        // overridden as not deprecated by fake driver
-        await driver.doubleClick();
-
-        await driver
-          .executeScript('fake: getDeprecatedCommandsCalled', [])
-          .should.eventually.eql(['callDeprecatedCommand', 'mobileShake']);
-      } finally {
-        await driver.deleteSession();
-      }
-    });
   });
 
   describe('Bidi protocol', function () {
     withServer();
     const capabilities = {...caps, webSocketUrl: true, 'appium:runClock': true};
+    /** @type import('webdriverio').Browser **/
     let driver;
 
     beforeEach(async function () {
@@ -601,6 +566,14 @@ describe('FakeDriver via HTTP', function () {
       await B.delay(750);
       collectedEvents.should.be.empty;
     });
+
+    it('should allow custom bidi commands', async function () {
+      let {result} = await driver.send({method: 'fake.getFakeThing', params: {}});
+      should.not.exist(result);
+      await driver.send({method: 'fake.setFakeThing', params: {thing: 'this is from bidi'}});
+      ({result} = await driver.send({method: 'fake.getFakeThing', params: {}}));
+      result.should.eql('this is from bidi');
+    });
   });
 });
 
@@ -633,15 +606,14 @@ describe('Bidi over SSL', function () {
   // since we update the FakeDriver.prototype below, make sure we update the FakeDriver which is
   // actually going to be required by Appium
   /** @type {import('@appium/types').DriverClass} */
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   let FakeDriver;
-  /** @type {string} */
-  let testServerBaseSessionUrl;
-  /** @type {import('sinon').SinonSandbox} */
-  let sandbox;
   let certPath = 'certificate.cert';
   let keyPath = 'certificate.key';
   const capabilities = {...caps, webSocketUrl: true};
   let should;
+  /** @type {string | undefined} */
+  let previousEnvValue;
 
   before(async function () {
     const chai = await import('chai');
@@ -657,7 +629,6 @@ describe('Bidi over SSL', function () {
       }
       return this.skip();
     }
-    sandbox = createSandbox();
     appiumHome = await tempDir.openDir();
     wdOpts.port = port = await getTestPort();
     testServerBaseUrl = `https://${TEST_HOST}:${port}`;
@@ -672,15 +643,20 @@ describe('Bidi over SSL', function () {
   });
 
   after(async function () {
-    await fs.rimraf(appiumHome);
-    await server.close();
+    if (server) {
+      await fs.rimraf(appiumHome);
+      await server.close();
+    }
   });
 
   beforeEach(async function () {
+    previousEnvValue = process.env.NODE_TLS_REJECT_UNAUTHORIZED;
+    process.env.NODE_TLS_REJECT_UNAUTHORIZED = '0';
     driver = await wdio({...wdOpts, protocol: 'https', strictSSL: false, capabilities});
   });
 
   afterEach(async function () {
+    process.env.NODE_TLS_REJECT_UNAUTHORIZED = previousEnvValue;
     if (driver) {
       await driver.deleteSession();
     }