+++ packages/base-driver/test/unit/express/static.spec.js
@@ -1,6 +1,5 @@
-// transpile:mocha
-
 import {welcome} from '../../../lib/express/static';
+// eslint-disable-next-line import/named
 import {createSandbox} from 'sinon';
 
 describe('welcome', function () {